+++ OpenC-1.0/scripts/verify_sh16_pe_runtime.py
@@ -45,6 +45,11 @@ EXPECTED_IMPORTS = [
     "GetProcAddress",
     "LocalFree",
     "FreeLibrary",
+    "CreatePipe",
+    "SetHandleInformation",
+    "CreateProcessW",
+    "WaitForSingleObject",
+    "GetExitCodeProcess",
 ]